feat: enhance SEO and polish UI components
- Added robots.txt and sitemap.xml
- Integrated external stats configuration
- Refactored skills renderer and UI components
- Improved page transitions and animations
- Cleaned up unused scroll-reveal modules
