+++ public/_headers
@@ -29,3 +29,15 @@
   Access-Control-Allow-Origin: *
   Access-Control-Allow-Methods: GET, POST, OPTIONS
   Access-Control-Allow-Headers: Content-Type
+
+# Cache static assets aggressively (1 year, immutable)
+/assets/*
+  Cache-Control: public, max-age=31536000, immutable
+
+# Cache HTML pages with revalidation (1 hour)
+/*.html
+  Cache-Control: public, max-age=3600, must-revalidate
+
+# Cache data files moderately (1 day)
+/data/*
+  Cache-Control: public, max-age=86400, must-revalidate
