+++ src/js/config/external-stats.js
@@ -1,0 +1,16 @@
+/**
+ * Manual stats for external projects not tracked in the main codebase
+ * (e.g., Timesheet Tracker, other side projects)
+ *
+ * These values are added to the skills.json counts to provide
+ * a more accurate representation of total lines written.
+ */
+export const externalStats = {
+  HTML: 200,
+  CSS: 500,
+  JavaScript: 10000,
+  TypeScript: 5000,
+  MySQL: 500,
+  React: 5000,
+  Python: 3500,
+};
