+++ src/pages/about.html
@@ -4,34 +4,69 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta name="theme-color" content="#0070f3" />
-    <meta name="description" content="Colin McArthur Developer Portfolio Site" />
+    <meta name="description" content="Learn about Colin McArthur - a self-taught full-stack developer from Peachland, BC. My journey, operating principles, and career highlights." />
     <link rel="stylesheet" href="../styles/tailwind.css" />
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link rel="preload" as="style" href="https://fonts.googleapis.com/css2?family=Inter:wght@400..900&family=Fira+Code:wght@400;700&display=swap" />
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Inter:wght@400..900&family=Fira+Code:wght@400;700&display=swap" media="print" onload="this.media='all'" />
     <noscript><link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Inter:wght@400..900&family=Fira+Code:wght@400;700&display=swap" /></noscript>
-    <title>Colin McArthur Developer Website</title>
-    <link rel="canonical" href="https://ColinMcArthur85.github.io/devsite/pages/about.html" />
-    <meta property="og:title" content="Colin McArthur Developer Website" />
-    <meta property="og:description" content="Colin McArthur Developer Portfolio Site" />
+    <title>About Colin McArthur | Self-Taught Developer</title>
+    <link rel="canonical" href="https://colinmcarthur85.github.io/devsite/pages/about.html" />
+    <meta property="og:title" content="About Colin McArthur | Self-Taught Developer" />
+    <meta property="og:description" content="Learn about Colin McArthur - a self-taught full-stack developer from Peachland, BC. My journey, operating principles, and career highlights." />
     <meta property="og:type" content="website" />
-    <meta property="og:url" content="https://ColinMcArthur85.github.io/devsite/pages/about.html" />
-    <meta property="og:image" content="https://ColinMcArthur85.github.io/devsite/assets/images/hero/hero-image-01.png" />
+    <meta property="og:url" content="https://colinmcarthur85.github.io/devsite/pages/about.html" />
+    <meta property="og:image" content="https://colinmcarthur85.github.io/devsite/assets/images/hero/hero-image-01.png" />
+    <meta property="og:site_name" content="Colin McArthur Developer Portfolio" />
+    <meta property="og:locale" content="en_US" />
     <meta name="twitter:card" content="summary_large_image" />
-    <meta name="twitter:title" content="Colin McArthur Developer Website" />
-    <meta name="twitter:description" content="Colin McArthur Developer Portfolio Site" />
-    <meta name="twitter:image" content="https://ColinMcArthur85.github.io/devsite/assets/images/hero/hero-image-01.png" />
+    <meta name="twitter:title" content="About Colin McArthur | Self-Taught Developer" />
+    <meta name="twitter:description" content="Learn about Colin McArthur - a self-taught full-stack developer from Peachland, BC." />
+    <meta name="twitter:image" content="https://colinmcarthur85.github.io/devsite/assets/images/hero/hero-image-01.png" />
+    <script type="application/ld+json">
+    {
+      "@context": "https://schema.org",
+      "@type": "WebPage",
+      "name": "About Colin McArthur",
+      "description": "Learn about Colin McArthur - a self-taught full-stack developer from Peachland, BC.",
+      "url": "https://colinmcarthur85.github.io/devsite/pages/about.html"
+    }
+    </script>
+    <script type="application/ld+json">
+    {
+      "@context": "https://schema.org",
+      "@type": "Person",
+      "name": "Colin McArthur",
+      "url": "https://colinmcarthur85.github.io/devsite/",
+      "jobTitle": "Full-Stack Developer",
+      "email": "colinsba@gmail.com",
+      "address": {
+        "@type": "PostalAddress",
+        "addressLocality": "Peachland",
+        "addressRegion": "BC",
+        "addressCountry": "Canada"
+      },
+      "sameAs": [
+        "https://github.com/ColinMcArthur85"
+      ]
+    }
+    </script>
   </head>
 
   <body class="dark:bg-dark-background bg-white font-sans text-gray-900 transition-colors duration-300 dark:text-white">
+    <!-- Skip to main content link for accessibility -->
+    <a href="#main-content" class="sr-only focus:not-sr-only focus:fixed focus:top-4 focus:left-4 focus:z-50 focus:bg-white focus:text-gray-900 focus:px-4 focus:py-2 focus:rounded focus:shadow-lg">Skip to main content</a>
+
     <div class="pointer-events-none fixed inset-0 -z-10 opacity-0 transition-opacity duration-500 dark:opacity-100">
       <div class="absolute inset-0 bg-linear-to-br from-primary/10 via-transparent to-secondary/10"></div>
     </div>
 
     <div id="menu-container" style="min-height: clamp(5.75rem, 8vw, 6.75rem); display: block"></div>
 
-    <section data-animate-on-scroll id="hero" class="hero-section relative overflow-hidden">
+    <!-- Main Content -->
+    <main id="main-content">
+    <section id="hero" class="hero-section relative overflow-hidden">
       <div class="absolute inset-0 -z-10">
         <div class="absolute -left-48 -top-32 h-112 w-md rounded-full bg-primary/20 blur-3xl dark:bg-primary/35"></div>
         <div class="absolute -right-56 -bottom-40 h-128 w-lg rounded-full bg-secondary/20 blur-3xl dark:bg-secondary/35"></div>
@@ -200,7 +235,7 @@
         <!-- Removed hero actions -->
       </div>
     </section>
-    <section data-animate-on-scroll id="story" class="relative isolate overflow-hidden py-24">
+    <section  id="story" class="relative isolate overflow-hidden py-24">
       <div class="absolute inset-0 -z-10">
         <div class="absolute -left-40 -top-24 h-104 w-104 rounded-full bg-primary/15 blur-3xl dark:bg-primary/30"></div>
         <div class="absolute -right-48 -bottom-32 h-112 w-md rounded-full bg-secondary/15 blur-3xl dark:bg-secondary/30"></div>
@@ -296,7 +331,7 @@
       </div>
     </section>
 
-    <section data-animate-on-scroll id="about" class="relative isolate overflow-hidden py-24">
+    <section  id="about" class="relative isolate overflow-hidden py-24">
       <div class="absolute inset-0 -z-10">
         <div class="absolute -left-40 -bottom-24 h-104 w-104 rounded-full bg-primary/15 blur-3xl dark:bg-primary/30"></div>
         <div class="absolute -right-48 -top-32 h-112 w-md rounded-full bg-secondary/15 blur-3xl dark:bg-secondary/30"></div>
@@ -372,7 +407,7 @@
       </div>
     </section>
 
-    <section data-animate-on-scroll id="timeline" class="relative isolate overflow-hidden py-24">
+    <section  id="timeline" class="relative isolate overflow-hidden py-24">
       <div class="absolute inset-0 -z-10">
         <div class="absolute left-1/2 top-0 h-88 w-88 -translate-x-1/2 rounded-full bg-primary/15 blur-3xl dark:bg-primary/25"></div>
         <div class="absolute -bottom-32 -right-24 h-96 w-96 rounded-full bg-secondary/15 blur-3xl dark:bg-secondary/25"></div>
@@ -452,6 +487,7 @@
         </div>
       </div>
     </section>
+    </main>
 
     <footer class="relative isolate overflow-hidden bg-slate-950 text-white">
       <div class="absolute inset-0 -z-10">
@@ -529,7 +565,6 @@
     <script type="module" src="../js/components/header.js"></script>
     <script type="module" src="../js/components/ui.js"></script>
     <script type="module" src="../js/components/boop.js"></script>
-    <script type="module" src="../js/modules/scroll-reveal.js"></script>
     <script type="module" src="../js/scripts.js"></script>
   </body>
 </html>
