+++ src/pages/skills.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta name="theme-color" content="#0070f3" />
-    <meta name="description" content="Colin McArthur — Developer Portfolio" />
+    <meta name="description" content="Explore Colin McArthur's technical skills - from HTML, CSS, and JavaScript to Python, MySQL, and AI-assisted workflows." />
     <link rel="stylesheet" href="../styles/tailwind.css" />
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
@@ -12,19 +12,38 @@
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Inter:wght@400..900&family=Fira+Code:wght@400;700&display=swap" media="print" onload="this.media='all'" />
     <noscript><link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Inter:wght@400..900&family=Fira+Code:wght@400;700&display=swap" /></noscript>
     <link rel="canonical" href="https://colinmcarthur85.github.io/devsite/pages/skills.html" />
-    <meta property="og:title" content="Colin McArthur — Skills" />
-    <meta property="og:description" content="A deeper look at my core competencies and the tools I wield." />
+    <meta property="og:title" content="Skills | Colin McArthur Developer" />
+    <meta property="og:description" content="Explore Colin McArthur's technical skills - from HTML, CSS, and JavaScript to Python, MySQL, and AI-assisted workflows." />
     <meta property="og:type" content="website" />
     <meta property="og:url" content="https://colinmcarthur85.github.io/devsite/pages/skills.html" />
     <meta property="og:image" content="https://colinmcarthur85.github.io/devsite/assets/images/hero/hero-image-01.png" />
+    <meta property="og:site_name" content="Colin McArthur Developer Portfolio" />
+    <meta property="og:locale" content="en_US" />
     <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:title" content="Skills | Colin McArthur Developer" />
+    <meta name="twitter:description" content="Explore Colin McArthur's technical skills and capabilities." />
+    <meta name="twitter:image" content="https://colinmcarthur85.github.io/devsite/assets/images/hero/hero-image-01.png" />
     <title>Skills | Colin McArthur</title>
+    <script type="application/ld+json">
+    {
+      "@context": "https://schema.org",
+      "@type": "WebPage",
+      "name": "Skills",
+      "description": "Explore Colin McArthur's technical skills - from HTML, CSS, and JavaScript to Python, MySQL, and AI-assisted workflows.",
+      "url": "https://colinmcarthur85.github.io/devsite/pages/skills.html"
+    }
+    </script>
   </head>
 
   <body class="bg-white font-sans text-gray-900 transition-colors duration-300 dark:bg-dark-background dark:text-white">
+    <!-- Skip to main content link for accessibility -->
+    <a href="#main-content" class="sr-only focus:not-sr-only focus:fixed focus:top-4 focus:left-4 focus:z-50 focus:bg-white focus:text-gray-900 focus:px-4 focus:py-2 focus:rounded focus:shadow-lg">Skip to main content</a>
+
     <div id="menu-container" style="min-height: clamp(5.75rem, 8vw, 6.75rem); display: block"></div>
 
-    <section data-animate-on-scroll id="hero" class="hero-section relative overflow-hidden">
+    <!-- Main Content -->
+    <main id="main-content">
+    <section id="hero" class="hero-section relative overflow-hidden">
       <div class="absolute inset-0 -z-10">
         <div class="absolute -left-48 -top-40 h-120 w-120 rounded-full bg-primary/20 blur-3xl dark:bg-primary/35"></div>
         <div class="absolute -right-48 -bottom-48 h-128 w-lg rounded-full bg-secondary/20 blur-3xl dark:bg-secondary/35"></div>
@@ -98,7 +117,7 @@
       </div>
     </section>
 
-    <section data-animate-on-scroll class="relative isolate overflow-hidden py-24">
+    <section  class="relative isolate overflow-hidden py-24">
       <div class="absolute inset-0 -z-10">
         <div class="absolute -left-48 -bottom-24 h-112 w-md rounded-full bg-primary/15 blur-3xl dark:bg-primary/30"></div>
         <div class="absolute -right-48 -top-32 h-120 w-120 rounded-full bg-secondary/15 blur-3xl dark:bg-secondary/30"></div>
@@ -303,7 +322,7 @@
       </div>
     </section>
 
-    <section data-animate-on-scroll class="relative isolate overflow-hidden py-24">
+    <section  class="relative isolate overflow-hidden py-24">
       <div class="absolute inset-0 -z-10">
         <div class="absolute left-1/2 top-0 h-80 w-80 -translate-x-1/2 rounded-full bg-primary/10 blur-3xl dark:bg-primary/20"></div>
         <div class="absolute -bottom-24 -right-32 h-88 w-88 rounded-full bg-secondary/10 blur-3xl dark:bg-secondary/20"></div>
@@ -406,7 +425,7 @@
                 Bootstrap 3 &mdash; intermediate
               </li>
               <li>
-                <span class="logo-slot boop" title="Froala logo"><img src="../assets/images/logos/froala_icon.png" alt="" /></span>
+                <span class="logo-slot boop" title="Froala logo"><img src="../assets/images/logos/froala_icon.png" alt="Froala logo" /></span>
                 Froala Editor &mdash; intermediate
               </li>
             </ul>
@@ -765,7 +784,7 @@
                   </li>
                   <li>
                     <span class="logo-slot boop" title="Namebright logo">
-                      <img src="../assets/images/logos/namebright.ico" alt="" />
+                      <img src="../assets/images/logos/namebright.ico" alt="NameBright logo" />
                     </span>
                     Namebright
                   </li>
@@ -1266,7 +1285,7 @@
             <h4 class="skill-heading">Email & Messaging</h4>
             <ul class="skill-list">
               <li>
-                <span class="logo-slot boop"><img src="../assets/images/logos/mailgun-icon.svg" alt="" /></span>
+                <span class="logo-slot boop"><img src="../assets/images/logos/mailgun-icon.svg" alt="Mailgun logo" /></span>
                 Mailgun
               </li>
               <li>
@@ -1300,6 +1319,7 @@
         </div>
       </div>
     </section>
+    </main>
 
     <footer class="relative isolate overflow-hidden bg-slate-950 text-white">
       <div class="absolute inset-0 -z-10">
@@ -1380,7 +1400,6 @@
     <script type="module" src="../js/components/boop.js"></script>
     <script type="module" src="../js/components/ui.js"></script>
     <script type="module" src="../js/skills.js"></script>
-    <script type="module" src="../js/modules/scroll-reveal.js"></script>
     <script type="module" src="../js/scripts.js"></script>
   </body>
 </html>
