+++ src/styles/tailwind/utils/_animations.css
@@ -61,33 +61,4 @@
     animation: boop 450ms ease-in-out;
   }
 
-  /* Scroll Reveal System */
-  /* If JS is enabled, hide elements until they are ready/visible to avoid flicker/shift */
-  :root.js [data-animate-on-scroll] {
-    opacity: 0;
-    transform: translateY(1.5rem);
-    pointer-events: none;
-  }
-
-  /* When element is scrolled into view or is already in viewport */
-  /* Must match :root.js specificity from the hiding rule above */
-  :root.js [data-animate-on-scroll].is-visible {
-    opacity: 1;
-    transform: translateY(0);
-    pointer-events: auto;
-  }
-
-  [data-animate-on-scroll].is-reveal-ready {
-    opacity: 0;
-    transform: translateY(1.5rem);
-    transition:
-      opacity 600ms cubic-bezier(0.2, 0.8, 0.2, 1),
-      transform 600ms cubic-bezier(0.2, 0.8, 0.2, 1);
-    transition-delay: var(--scroll-reveal-delay, 0ms);
-  }
-
-  [data-animate-on-scroll].is-reveal-ready.is-visible {
-    opacity: 1;
-    transform: translateY(0);
-  }
 }
