+++ tailwind.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  content: ["./src/**/*.{html,js}", "./public/**/*.{html,js,php}", "./src/*.html", "./*.php"],
+  content: ["./src/**/*.{html,js}", "./public/**/*.{html,js}", "./src/*.html"],
   darkMode: "class",
   theme: {},
   plugins: [],
