+++ vite.config.mjs
@@ -17,7 +17,7 @@ export default defineConfig({
         contact: resolve(__dirname, "src/pages/contact.html"),
         skills: resolve(__dirname, "src/pages/skills.html"),
         "projects/api/image_search/index": resolve(__dirname, "src/projects/api/image_search/index.html"),
-        // showcase: resolve(__dirname, 'src/pages/showcase.html'), // Uncomment if exists
+        showcase: resolve(__dirname, "src/pages/showcase.html"),
       },
     },
   },
